Using BigDecimal instead of double

+++ src/main/java/com/mauriciotogneri/botcoin/provider/Price.java
@@ -1,11 +1,13 @@
 package com.mauriciotogneri.botcoin.provider;
 
+import java.math.BigDecimal;
+
 public class Price
 {
     public final long timestamp;
-    public final double value;
+    public final BigDecimal value;
 
-    public Price(long timestamp, double value)
+    public Price(long timestamp, BigDecimal value)
     {
         this.timestamp = timestamp;
         this.value = value;
